Move mirror schemas to namespace api folder
